Created folders, moved logger, added an exclude from code attribute
