+++ InternProject.CsvFileConverter.Tests/Core.Conversions.Tests/DataExtractorTest.cs
