+++ InternProject.CsvFileConverter/Autofac/IocBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics.CodeAnalysis;
 using Autofac;
 using CsvFileConverter.Logging;
 using CsvFileConverter.MainProgramme;
@@ -10,30 +11,31 @@ using Serilog.Exceptions;
 
 namespace CsvFileConverter
 {
+    [ExcludeFromCodeCoverage]
     public class IocBuilder
-    {
-        public static IContainer Build()
         {
-            var logger = Log.Logger = new LoggerConfigFile().SeriLogConfig;
+            public static IContainer Build()
+            {
+                var logger = Log.Logger = new LoggerConfigFile().SeriLogConfig;
 
-            var builder = new ContainerBuilder();
+                var builder = new ContainerBuilder();
 
-            builder.RegisterType<CsvToJsonConverter>();
-            builder.RegisterInstance(logger);
-            builder.RegisterType<FileReader>().As<IFileReader>();
-            builder.RegisterType<DataExtractor>().As<IDataExtractor>();
-            builder.RegisterType<JsonTextFormatter>().As<ITextFormatter>();
-            builder.RegisterType<XmlTextFormatter>().As<ITextFormatter>();
-            builder.RegisterType<FileWriter>().As<IFileWriter>();
-            
-            builder.RegisterType<IntFieldValidator>().As<IFieldValidator>();
-            builder.RegisterType<DoubleFieldValidator>().As<IFieldValidator>();
-            builder.RegisterType<DateFieldValidator>().As<IFieldValidator>();
-            builder.RegisterType<StringFieldValidator>().As<IFieldValidator>();
-            builder.RegisterTypes(typeof(DealDataRawValidator)).As<IValidator<DealDataRaw>>();
+                builder.RegisterType<CsvToJsonConverter>();
+                builder.RegisterInstance(logger);
+                builder.RegisterType<FileReader>().As<IFileReader>();
+                builder.RegisterType<DataExtractor>().As<IDataExtractor>();
+                builder.RegisterType<JsonTextFormatter>().As<ITextFormatter>();
+                builder.RegisterType<XmlTextFormatter>().As<ITextFormatter>();
+                builder.RegisterType<FileWriter>().As<IFileWriter>();
 
-            var container= builder.Build();
-            return container;
+                builder.RegisterType<IntFieldValidator>().As<IFieldValidator>();
+                builder.RegisterType<DoubleFieldValidator>().As<IFieldValidator>();
+                builder.RegisterType<DateFieldValidator>().As<IFieldValidator>();
+                builder.RegisterType<StringFieldValidator>().As<IFieldValidator>();
+
+                var container = builder.Build();
+                return container;
+            }
         }
     }
-}
+
