+++ InternProject.CsvFileConverter/CsvFileConverter.csproj
@@ -14,10 +14,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Autofac" Version="4.6.2" />
+    <PackageReference Include="Autofac" Version="4.8.0" />
     <PackageReference Include="CsvHelper" Version="7.1.0" />
-    <PackageReference Include="FluentAssertions" Version="5.2.0" />
-    <PackageReference Include="FluentValidation" Version="7.5.2" />
+    <PackageReference Include="FluentAssertions" Version="5.3.0" />
+    <PackageReference Include="FluentValidation" Version="7.6.0-preview1" />
     <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
     <PackageReference Include="NHibernate.Validator" Version="5.1.0" />
     <PackageReference Include="NSubstitute" Version="3.1.0" />
@@ -26,6 +26,10 @@
     <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
     <PackageReference Include="Serilog.Sinks.File" Version="4.0.1-dev-00795" />
     <PackageReference Include="System.Runtime.InteropServices.RuntimeInformation" Version="4.3.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Files\" />
   </ItemGroup>
 
 </Project>
