+++ InternProject.CsvFileConverter/Extensions/Formatters/JsonTextFormatter.cs
@@ -1,10 +1,12 @@
 ﻿using System;
+using System.Diagnostics.CodeAnalysis;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using NHibernate.Linq;
 
 namespace CsvFileConverter
 {
+    [ExcludeFromCodeCoverage]
     public class JsonTextFormatter : ITextFormatter
     {
         public FormatterType Type => FormatterType.Json;
