--- InternProject.CsvFileConverter/InputOutputFiles/Deal.json
@@ -1,20 +1,0 @@
-[
-  {
-    "V3DealID": "02B4EFADE6",
-    "eFrontDealID": "02B4EFADE60B48339D13F93EB851943C",
-    "DealName": "Marston (Project Magenta)",
-    "V3CompanyID": "",
-    "V3CompanyName": "",
-    "SectorID": "",
-    "Sector": "",
-    "CountryID": "229",
-    "Country": "United Kingdom",
-    "TransactionTypeID": "",
-    "Transaction Type": "",
-    "Transaction Fees": "",
-    "Other Fees": "",
-    "Currency": "EUR",
-    "ActiveInActive": "Active",
-    "ExitDate": ""
-  }
-]
