+++ InternProject.CsvFileConverter/Interfaces/Core.Conversions.Interfaces/IDataExtractor.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System.Diagnostics.CodeAnalysis;
+using System.IO;
 
 namespace CsvFileConverter
 {
