+++ InternProject.CsvFileConverter/Interfaces/Core.IO.Interfaces/IFileReader.cs
