+++ InternProject.CsvFileConverter/Interfaces/Extensions.Interfaces/ITextFormatter.cs
