+++ InternProject.CsvFileConverter/Logging/LoggerConfigFile.cs
@@ -1,11 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Text;
 using Serilog;
 using Serilog.Exceptions;
 
 namespace CsvFileConverter.Logging
 {
+    [ExcludeFromCodeCoverage]
     class LoggerConfigFile 
     {
