+++ InternProject.CsvFileConverter/Logging/LoggerEnrichmentConfigExt.cs
@@ -1,8 +1,10 @@
-﻿using Serilog;
+﻿using System.Diagnostics.CodeAnalysis;
+using Serilog;
 using Serilog.Configuration;
 
 namespace CsvFileConverter
 {
+    [ExcludeFromCodeCoverage]
     public static class LoggerEnrichmentConfigExt
     {
         public static LoggerConfiguration WithCaller(this LoggerEnrichmentConfiguration enrichmentConfiguration)
