+++ InternProject.CsvFileConverter/Validation/DoubleFieldValidator.cs
@@ -1,7 +1,9 @@
 ﻿using System;
+using System.Diagnostics.CodeAnalysis;
 
 namespace CsvFileConverter
 {
+    [ExcludeFromCodeCoverage]
     public class DoubleFieldValidator : IFieldValidator
     {
         public Type TypeToValidate => typeof(double);
