+++ InternProject.CsvFileConverter/Validation/StringFieldValidator.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 using System.Data;
+using System.Diagnostics.CodeAnalysis;
 using Serilog;
 
 namespace CsvFileConverter
 {
+    [ExcludeFromCodeCoverage]
     public class StringFieldValidator : IFieldValidator
     {
         public Type TypeToValidate => typeof(string);
